pyproject.toml
[tool.poetry]
name = "synthdetect"
version = "1.0.0-alpha"
description = "AI-Generated Text Detection via Disrupt-and-Recover and Fine-Grained Attribution"
authors = ["SynthDetect Team"]
license = "MIT"
readme = "README.md"
keywords = ["ai-detection", "nlp", "text-analysis", "machine-learning"]

[tool.poetry.dependencies]
python = "^3.10"
torch = {version = "^2.1.0", source = "pytorch-cpu"}
transformers = "^4.35.0"
sentence-transformers = "^2.2.2"
spacy = "^3.7.0"
google-genai = "^1.0.0"
faiss-cpu = "^1.7.4"
lightgbm = "^4.1.0"
pydantic = "^2.5.0"
pydantic-settings = "^2.1.0"
numpy = "^1.26.0"
scikit-learn = "^1.3.0"
pandas = "^2.1.0"
python-dotenv = "^1.0.0"
httpx = "^0.25.0"
pyyaml = "^6.0.1"
tqdm = "^4.66.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
pytest-asyncio = "^0.21.1"
pytest-cov = "^4.1.0"
black = "^23.11.0"
isort = "^5.12.0"
flake8 = "^6.1.0"
mypy = "^1.7.0"
ipython = "^8.17.2"

[tool.poetry.group.research.dependencies]
jupyter = "^1.0.0"
matplotlib = "^3.8.0"
seaborn = "^0.13.0"
plotly = "^5.18.0"
shap = "^0.43.0"
optuna = "^3.4.0"
datasets = "^2.15.0"

[[tool.poetry.source]]
name = "pytorch-cpu"
url = "https://download.pytorch.org/whl/cpu"
priority = "explicit"

[tool.black]
line-length = 100
target-version = ['py310']
include = '\.pyi?$'

[tool.isort]
profile = "black"
line_length = 100

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks integration tests requiring external services",
]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
